+++ src/bp/scripts/features/index.ts
@@ -3,6 +3,7 @@ import "./config";
 
 import "./player_loop";
 
+import "./blast_door";
 import "./doors";
 import "./fnet";
 import "./furniture";