+++ src/bp/scripts/logic/blastDoor/index.ts
@@ -1,2 +1,0 @@
-import "./dummyBlock";
-import "./preview";