+++ CHANGELOG.md
@@ -29,6 +29,7 @@ _Most_ notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- Blast Door creation form is not displayed properly.
 - Toilets and Chairs are not sittable.
 - Dash Block does not work.